refactor(realm): extract fs bridge from js-realm-shared
